# Conflicts:
#	installer.nsi
#	resources/web/flutter_web/flutter_bootstrap.js
#	resources/web/flutter_web/flutter_service_worker.js
#	resources/web/flutter_web/main.dart.js
#	resources/web/flutter_web/version.changelog
#	resources/web/flutter_web/version.json
#	scripts/flatpak/io.github.Snapmaker.Snapmaker_Orca.metainfo.xml
#	src/common_func/common_func.hpp
#	src/slic3r/GUI/AboutDialog.cpp
#	src/slic3r/GUI/FilamentColorDialog.cpp
#	src/slic3r/GUI/GUI_App.cpp
#	src/slic3r/GUI/MainFrame.cpp
#	src/slic3r/GUI/NetworkTestDialog.cpp
#	src/slic3r/GUI/Plater.cpp
#	src/slic3r/GUI/Tab.cpp
#	src/slic3r/GUI/Widgets/SegmentedToggle.cpp
#	src/slic3r/GUI/filamentsync/FilamentColorMapBox.cpp
#	src/slic3r/GUI/filamentsync/FilamentColorMapBoxGroup.cpp
#	src/slic3r/GUI/filamentsync/FilamentData.hpp
#	src/slic3r/GUI/filamentsync/FilamentScrollBar.cpp
#	src/slic3r/GUI/filamentsync/MachineFilamentPicker.cpp
#	src/slic3r/GUI/filamentsync/MachineFilamentPicker.hpp
#	src/slic3r/GUI/filamentsync/SyncFilamentColorDialog.cpp
#	version.inc

+++ src/slic3r/GUI/GUI_App.cpp
@@ -2257,6 +2257,7 @@ void GUI_App::init_app_config()
     }
     MixedFilamentManager::set_auto_generate_enabled(app_config->get_bool("auto_generate_gradients"));
     set_logging_level(Slic3r::level_string_to_boost(app_config->get("log_severity_level")));
+
 }
 
 // returns true if found newer version and user agreed to use it
@@ -2342,7 +2343,7 @@ void GUI_App::do_notify_flutter_web_copy_failure()
         }
         break;
     default: 
-        BOOST_LOG_TRIVIAL(error) << "FlutterWebCopyStatus other status" << m_flutter_web_copy_status;
+        BOOST_LOG_TRIVIAL(error) << "FlutterWebCopyStatus other status" << static_cast<int>(m_flutter_web_copy_status);
         break;
     }
 }
